feat: rename Article component to PageArticle (#83)
